Implemented interior and exterior light segregation
When a building is sliced; interior and exterior light sources bleed into the others domain. However, the light sources don't bleed into eachothers domain when the building isn't being sliced.  Inorder, to ensure that the sliced building is similar to the non-sliced version; a invisible shadow castable object has been added to the building prefab. The invisible shadow castable poccess the same mesh as the building(without any slicing). As, a result interior and exterior light sources can be segregated.